chore(storage-browser): remove duplicate FileItem interface, cleanup type and util locations
