+++ packages/react-storage/src/components/StorageBrowser/actions/index.ts
@@ -5,7 +5,6 @@ export {
   CopyHandlerData,
   CopyHandlerInput,
   CopyHandlerOutput,
-  createFileDataItem,
   createFolderHandler,
   CreateFolderHandler,
   CreateFolderHandlerData,
@@ -25,11 +24,8 @@ export {
   DeleteHandlerOptions,
   DeleteHandlerOutput,
   FileData,
-  FileDataItem,
-  FileItem,
   FolderData,
-  isFileDataItem,
-  isFileItem,
+  getFileKey,
   ListHandler,
   ListHandlerInput,
   ListHandlerOptions,
