+++ packages/react-storage/src/components/StorageBrowser/displayText/index.ts
@@ -1,6 +1,7 @@
 export { DisplayTextProvider, useDisplayText } from './context';
 export { DEFAULT_STORAGE_BROWSER_DISPLAY_TEXT as DefaultStorageBrowserDisplayText } from './libraries';
 export {
+  DefaultActionViewDisplayText,
   LocationDetailViewDisplayText,
   LocationsViewDisplayText,
   CreateFolderViewDisplayText,
