+++ packages/react-storage/src/components/StorageBrowser/providers/index.ts
@@ -4,8 +4,15 @@ export {
   RegisterAuthListener,
 } from './configuration';
 export {
+  createFileDataItem,
+  FileDataItem,
   FileItem,
   FileItems,
-  StoreProviderProps,
+  isFileDataItem,
+  isFileItem,
+  LocationState,
   StoreProvider,
+  StoreProviderProps,
+  useStore,
+  UseStoreState,
 } from './store';
