+++ packages/react-storage/src/components/StorageBrowser/providers/store/files/index.ts
@@ -1,2 +1,3 @@
 export { FilesProvider, FilesContext, useFiles } from './context';
 export * from './types';
+export { isFileItem } from './utils';
