+++ packages/react-storage/src/components/StorageBrowser/providers/store/files/utils.ts
@@ -71,3 +71,6 @@ export const parseFileSelectParams = (
   const [selectType, ...rest] = value;
   return [selectType, !rest?.length ? undefined : { accept: rest.join() }];
 };
+
+export const isFileItem = (value: unknown): value is FileItem =>
+  !!(value as FileItem).file;
