+++ packages/react-storage/src/components/StorageBrowser/providers/store/index.ts
@@ -1,3 +1,9 @@
-export { FileItem, FileItems } from './files';
+export { FileItem, FileItems, isFileItem } from './files';
+export { LocationState } from './location';
+export {
+  createFileDataItem,
+  FileDataItem,
+  isFileDataItem,
+} from './locationItems';
 export { StoreProvider, StoreProviderProps } from './StoreProvider';
 export { useStore, UseStoreState } from './useStore';
