+++ packages/react-storage/src/components/StorageBrowser/providers/store/locationItems/types.ts
@@ -1,0 +1,5 @@
+import { FileData, TaskData } from '../../../actions';
+
+export interface FileDataItem extends FileData, TaskData {
+  fileKey: string;
+}
